Robert

+++ rocket.h
@@ -21,7 +21,7 @@ namespace rocketNS
 	const int HEIGHT = 29;                  // image height
 	const int X = GAME_WIDTH / 2 - WIDTH / 2;   // location on screen
 	const int Y = GAME_HEIGHT / 2 - HEIGHT / 2;
-	const float ROTATION_RATE = (float)1.3 * PI; // radians per second
+	const float ROTATION_RATE = 1.3 * PI; // radians per second
 	const float SPEED = 300;                // 100 pixels per second
 	const float MASS = 400.0f;              // mass
 	const int   TEXTURE_COLS = 1;           // texture has 8 columns
